Fix put route for product tags

+++ routes/api/product-routes.js
@@ -6,7 +6,6 @@ const { Product, Category, Tag, ProductTag } = require('../../models');
 // get all products
 router.get('/', async (req, res) => {
   // find all products
-  // be sure to include its associated Category and Tag data
   try{
     const products = await Product.findAll({include:[Category,Tag]})
     res.status(200).json(products)
@@ -20,9 +19,8 @@ router.get('/', async (req, res) => {
 // get one product
 router.get('/:id', async (req, res) => {
   // find a single product by its `id`
-  // be sure to include its associated Category and Tag data
   try{
-    const selectedProduct = await Product.findOne({where:{id:req.params.id},include:[Category,Tag]})
+    const selectedProduct = await Product.findByPk(req.params.id,{include:[Category,Tag]})
     if(!selectedProduct){
       res.status(400).json('No product in database matching that id')
       return
@@ -36,14 +34,8 @@ router.get('/:id', async (req, res) => {
 
 // create new product
 router.post('/', (req, res) => {
-  /* req.body should look like this...
-    {
-      product_name: "Basketball",
-      price: 200.00,
-      stock: 3,
-      tagIds: [1, 2, 3, 4]
-    }
-  */
+  
+  // If not all information is passed in
   if (!req.body.productName || !req.body.price || !req.body.stock || !req.body.tagIds || !req.body.categoryId){
     res.status(400).json('Not enough information to create a new product. ');
     return
@@ -65,67 +57,90 @@ router.post('/', (req, res) => {
     })
     .then((productTagIds) => res.status(200).json(productTagIds))
     .catch((err) => {
-      console.log(err);
       res.status(500).json(err);
     });
 });
 
 // update product
 router.put('/:id', (req, res) => {
+
   // update product data
-  console.log('put',req.body)
   Product.update(req.body, {
     where: {
       id: req.params.id,
     },
   })
     .then((product) => {
+      
+      // If no new information is entered
+      if(product[0]===0&&!req.body.tagId){
+        return 'No product was updated, please check information provided'
+      }
+      
       // find all associated tags from ProductTag
       return ProductTag.findAll({ where: { productId: req.params.id } });
     })
     .then((productTags) => {
-      // get list of current tag_ids
-      console.log(productTags)
+
+      // If no new information was entered and no tagIds provided
+      if (productTags === 'No product was updated, please check information provided'){
+
+        return {status:400,msg:'No product was updated, please check information provided'}
+      }
+
+      // If no tagId array is present, then the information is updated already
+      if (!req.body.tagId){
+        return {status:200,msg:'Product updated'}
+      }
+
+      // get list of current tagIds
       const productTagIds = productTags.map(({ tagId }) => tagId);
+      
+      // If no new information entered and tagId array is the same as current tagId array
+      if (JSON.stringify(req.body.tagId.sort())==JSON.stringify(productTagIds)){
+        return {status:400,msg:'Product not updated, no new information added'}
+      }
+
       // create filtered list of new tag_ids
-      console.log(productTagIds,req.body.tagId)
       const newProductTags = req.body.tagId
         .filter((tagId) => !productTagIds.includes(tagId))
         .map((tagId) => {
-          console.log(tagId,'INSIDE MAP')
           return {
             productId: req.params.id,
             tagId,
           };
         });
-        console.log(newProductTags,'NEW TAGS    AS SDDASADSD A')
+
       // figure out which ones to remove
       const productTagsToRemove = productTags
         .filter(({ tagId }) => !req.body.tagId.includes(tagId))
         .map(({ id }) => id);
-        console.log(productTagsToRemove,'test')
+
       // run both actions
-      return Promise.all([
+      Promise.all([
         ProductTag.destroy({ where: { id: productTagsToRemove } }),
         ProductTag.bulkCreate(newProductTags),
       ]);
+
+      // Once product tags are updated
+      return {status:200,msg:'Product updated with appropriate tag changes'}
     })
-    .then((updatedProductTags) => res.json(updatedProductTags))
+    .then((response) => res.status(response.status).json(response.msg)) //response object used to prevent multiple res.json calls
     .catch((err) => {
-      // console.log(err);
-      res.status(400).json(err);
+      res.status(500).json(err);
     });
 });
 
 router.delete('/:id', async (req, res) => {
   // delete one product by its `id` value
   try{
-    
     const deletedProduct = await Product.destroy({
       where:{
         id:req.params.id
       }
     })
+
+    // If no rows deleted
     if (!deletedProduct){
       res.status(400).json('No matching product in the database')
       return
